pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "langfuse>=3.14.1",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "rich>=14.3.2",
    "sentence-transformers>=5.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
